+++ src/kanban/node_layout.rs
@@ -1,6 +1,5 @@
 use std::cmp::Ordering;
 
-use std::ops::{Index, IndexMut};
 use std::time::Instant;
 
 use super::*;